requirements.txt
pyglet~=1.5.28
dicom~=0.9.9.post1
numpy~=1.24.4
Pillow~=10.1.0